+++ web/src/pages/flow/canvas/node/email-node.tsx
@@ -1,8 +1,8 @@
+import { IEmailNode } from '@/interfaces/database/flow';
 import { Handle, NodeProps, Position } from '@xyflow/react';
 import { Flex } from 'antd';
 import classNames from 'classnames';
 import { useState } from 'react';
-import { IEmailNode } from '../../interface';
 import { LeftHandleStyle, RightHandleStyle } from './handle-icon';
 import styles from './index.less';
 import NodeHeader from './node-header';
