+++ web/src/pages/flow/form/generate-form/dynamic-parameters.tsx
@@ -1,15 +1,14 @@
 import { EditableCell, EditableRow } from '@/components/editable-cell';
 import { useTranslate } from '@/hooks/common-hooks';
 import { DeleteOutlined } from '@ant-design/icons';
-import { Node } from '@xyflow/react';
 import { Button, Flex, Select, Table, TableProps } from 'antd';
 import { useBuildComponentIdSelectOptions } from '../../hooks/use-get-begin-query';
-import { IGenerateParameter, NodeData } from '../../interface';
+import { IGenerateParameter, RAGFlowNodeType } from '../../interface';
 import { useHandleOperateParameters } from './hooks';
 
 import styles from './index.less';
 interface IProps {
-  node?: Node<NodeData>;
+  node?: RAGFlowNodeType;
 }
 
 const components = {
