+++ web/src/pages/flow/interface.ts
@@ -1,49 +1,11 @@
-import { BuiltInNode, Node } from '@xyflow/react';
+import { RAGFlowNodeType } from '@/interfaces/database/flow';
 import { FormInstance } from 'antd';
-
-export interface DSLComponentList {
-  id: string;
-  name: string;
-}
 
 export interface IOperatorForm {
   onValuesChange?(changedValues: any, values: any): void;
   form?: FormInstance;
-  node?: Node<NodeData>;
+  node?: RAGFlowNodeType;
   nodeId?: string;
-}
-
-export interface IBeginForm {
-  prologue?: string;
-}
-
-export interface IRetrievalForm {
-  similarity_threshold?: number;
-  keywords_similarity_weight?: number;
-  top_n?: number;
-  top_k?: number;
-  rerank_id?: string;
-  empty_response?: string;
-  kb_ids: string[];
-}
-
-export interface IGenerateForm {
-  max_tokens?: number;
-  temperature?: number;
-  top_p?: number;
-  presence_penalty?: number;
-  frequency_penalty?: number;
-  cite?: boolean;
-  prompt: number;
-  llm_id: string;
-  parameters: { key: string; component_id: string };
-}
-export interface ICategorizeItem {
-  name: string;
-  description?: string;
-  examples?: string;
-  to?: string;
-  index: number;
 }
 
 export interface IGenerateParameter {
@@ -55,83 +17,6 @@ export interface IGenerateParameter {
 export interface IInvokeVariable extends IGenerateParameter {
   value?: string;
 }
-
-export type ICategorizeItemResult = Record<
-  string,
-  Omit<ICategorizeItem, 'name'>
->;
-export interface ICategorizeForm extends IGenerateForm {
-  category_description: ICategorizeItemResult;
-}
-
-export interface IRelevantForm extends IGenerateForm {
-  yes: string;
-  no: string;
-}
-
-export interface ISwitchCondition {
-  items: ISwitchItem[];
-  logical_operator: string;
-  to: string;
-}
-
-export interface ISwitchItem {
-  cpn_id: string;
-  operator: string;
-  value: string;
-}
-
-export interface ISwitchForm {
-  conditions: ISwitchCondition[];
-  end_cpn_id: string;
-  no: string;
-}
-
-export type BaseNodeData = {
-  label: string; // operator type
-  name: string; // operator name
-  color?: string;
-};
-
-export type BaseNode<T = any> = Node<{ form: T } & BaseNodeData>;
-
-export type IBeginNode = BaseNode<IBeginForm>;
-export type IRetrievalNode = BaseNode<IRetrievalForm>;
-export type IGenerateNode = BaseNode<IGenerateForm>;
-export type ICategorizeNode = BaseNode<ICategorizeForm>;
-export type ISwitchNode = BaseNode<ISwitchForm>;
-export type IRagNode = BaseNode;
-export type IRelevantNode = BaseNode;
-export type ILogicNode = BaseNode;
-export type INoteNode = BaseNode;
-export type IMessageNode = BaseNode;
-export type IRewriteNode = BaseNode;
-export type IInvokeNode = BaseNode;
-export type ITemplateNode = BaseNode;
-export type IEmailNode = BaseNode;
-export type IIterationNode = BaseNode;
-export type IIterationStartNode = BaseNode;
-export type IKeywordNode = BaseNode;
-
-export type RAGFlowNodeType =
-  | BuiltInNode
-  | IBeginNode
-  | IRetrievalNode
-  | IGenerateNode
-  | ICategorizeNode
-  | ISwitchNode;
-
-export type NodeData = {
-  label: string; // operator type
-  name: string; // operator name
-  color?: string;
-  form:
-    | IBeginForm
-    | IRetrievalForm
-    | IGenerateForm
-    | ICategorizeForm
-    | ISwitchForm;
-};
 
 export type IPosition = { top: number; right: number; idx: number };
